+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 New:
 - Added support for DataSets
 - Added new Device rename command
-- Redesigned UI for Broker configuration
+- Redesigned Broker configuration UI
+- Added support for manual connection of the EoN
+- Added connect command for EoN
+- Added rename command for EoN node
 
 ### 1.4.1: Maintenance Release